NEW: Funcion para extraer en otro Vendor
